backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "transcript-editor-backend"
version = "0.1.0"
description = "Backend for the transcript editor application"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "python-multipart==0.0.6",
    "google-generativeai==0.3.2",
    "openai==1.12.0",
    "python-dotenv==1.0.0",
    "sqlalchemy==2.0.23",
    "pydantic==2.5.2",
    "httpx==0.25.2",
    "python-docx==1.1.0",
    "alembic==1.13.1",
    "psycopg2-binary",
    "pydantic-settings>=2.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
]

[tool.setuptools]
# Optional: Add package discovery settings if your project structure requires it
# package-dir = {"" = "src"}
# packages = find:"src"

[tool.pytest.ini_options]
pythonpath = [".", "app"]
# Add other pytest configurations if needed